+++ src/js/language/be_NL.js
@@ -184,6 +184,9 @@
         imei: {
             'default': 'Geef een geldig IMEI nummer'
         },
+        imo: {
+            'default': 'Geef een geldig IMO nummer'
+        },
         integer: {
             'default': 'Geef een geldig nummer'
         },